exynos_mem: an alternative temporary workaround for exynos-mem security hole
This assumption may be wrong but I guess exynos-mem is meant only for reading from/writing to CMA regions (for camera firmware for example) and limiting the access to those regions may prevent the security hole while not breaking any functionality until Samsung fixes this in a more proper way.

Keep in mind that this modification is only for S2 (i9100) and S3 (i9300). You will have to append dma reservations from exynos_reserve function in the corresponding platform file in /arch/arm/mach-exynos4 folder.